Split the config files for each package
