+++ packages/beacon-node/package.json
@@ -77,10 +77,10 @@
     "lint:fix": "yarn run lint --fix",
     "pretest": "yarn run check-types",
     "test": "yarn test:unit && yarn test:e2e",
-    "test:unit:minimal": "vitest --run --segfaultRetry 3 --dir test/unit/ --coverage",
+    "test:unit:minimal": "vitest --run --segfaultRetry 3 --dir test/unit/",
     "test:unit:mainnet": "LODESTAR_PRESET=mainnet vitest --run --dir test/unit-mainnet",
     "test:unit": "wrapper() { yarn test:unit:minimal $@ && yarn test:unit:mainnet $@; }; wrapper",
-    "test:e2e": "LODESTAR_PRESET=minimal vitest --run --segfaultRetry 3 --poolOptions.threads.singleThread true --dir test/e2e",
+    "test:e2e": "LODESTAR_PRESET=minimal vitest --run --segfaultRetry 3 --config vitest.config.e2e.ts --dir test/e2e",
     "test:sim": "vitest --run test/sim/**/*.test.ts",
     "test:sim:merge-interop": "vitest --run test/sim/merge-interop.test.ts",
     "test:sim:mergemock": "vitest --run test/sim/mergemock.test.ts",
