+++ packages/beacon-node/vitest.config.spec.ts
@@ -1,19 +1,11 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import vitestConfig from "../../vitest.base.config";
+import vitestConfig from "../../vitest.base.spec.config";
 
 export default mergeConfig(
   vitestConfig,
   defineConfig({
     test: {
       globalSetup: ["./test/globalSetup.ts"],
-      testTimeout: 60_000,
-      passWithNoTests: true,
-      pool: "threads",
-      poolOptions: {
-        threads: {
-          isolate: false,
-        },
-      },
     },
   })
 );
