+++ packages/cli/package.json
@@ -31,8 +31,8 @@
     "lint": "eslint --color --ext .ts src/ test/",
     "lint:fix": "yarn run lint --fix",
     "pretest": "yarn run check-types",
-    "test:unit": "vitest --run --dir test/unit/ --coverage",
-    "test:e2e": "vitest --run --poolOptions.threads.singleThread true --dir test/e2e/",
+    "test:unit": "vitest --run --dir test/unit/",
+    "test:e2e": "vitest --run --config vitest.config.e2e.ts --dir test/e2e/",
     "test:sim:multifork": "LODESTAR_PRESET=minimal node --loader ts-node/esm test/sim/multi_fork.test.ts",
     "test:sim:mixedclient": "LODESTAR_PRESET=minimal node --loader ts-node/esm test/sim/mixed_client.test.ts",
     "test:sim:endpoints": "LODESTAR_PRESET=minimal node --loader ts-node/esm test/sim/endpoints.test.ts",
