+++ packages/logger/vitest.config.e2e.ts
@@ -1,5 +1,5 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import vitestConfig from "../../vitest.base.unit.config";
+import vitestConfig from "../../vitest.base.e2e.config";
 
 export default mergeConfig(
   vitestConfig,
