+++ packages/spec-test-util/package.json
@@ -47,8 +47,8 @@
     "lint": "eslint --color --ext .ts src/ test/",
     "lint:fix": "yarn run lint --fix",
     "pretest": "yarn run check-types",
-    "test:unit": "vitest --run --passWithNoTests --dir test/unit/ --coverage",
-    "test:e2e": "vitest --run --dir test/e2e/",
+    "test:unit": "vitest --run --passWithNoTests --dir test/unit/",
+    "test:e2e": "vitest --run --config vitest.config.e2e.ts --dir test/e2e/",
     "check-readme": "typescript-docs-verifier"
   },
   "repository": {
