+++ vitest.base.spec.config.ts
@@ -1,12 +1,15 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import vitestConfig from "../../vitest.base.unit.config";
+import vitestConfig from "./vitest.base.unit.config.js";
 
 export default mergeConfig(
   vitestConfig,
   defineConfig({
     test: {
-      globalSetup: ["./test/globalSetup.ts"],
+      coverage: {
+        enabled: false,
+      },
       testTimeout: 60_000,
+      hookTimeout: 60_000,
       passWithNoTests: true,
       pool: "threads",
       poolOptions: {
